Replace auto-merge action to fastify/github-action-merge-dependabot
I replaced `ahmadnassri/action-dependabot-auto-merge` with another one because the previous one uses the command `@dependabot/merge` which can't be posted by the GitHub Actions bot. So, it requires a personal token to be used, which isn't a good idea for enterprise projects due to variable team composition. Unlike that, the new action uses the GitHub Auto-merge feature, which requires only the default GITHUB_TOKEN and enabling the Auth-merge feature in settings.