Fix the wrong revision for the `test` workflow checkout
The issue appeared because `pull_request_target` uses the PR base branch as `GITHUB_REF`: https://github.com/actions/checkout/issues/518. In the issue, they offer to use a merged revision implicitly. But that enables stealing secrets from the repo: https://blog.gitguardian.com/github-actions-security-cheat-sheet/#be-extra-careful-with-the-pullrequesttarget-trigger.

Currently, the `pull_request_target` trigger is used in the `test` workflow to enable adding a coverage report comment to the PR based on fork repos.

There is a solution in GitHub docs to leave a comment, but it seems to be an over-engineering: https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#using-data-from-the-triggering-workflow. The easiest way just to disable creating the comment for PRs from forks.