RDMP-423: Onboard Backstage Catalog File (#208)
## Background
Part of [RDMP-423]

This pull request is about updates or creation of your component's
catalog-info.yaml file using component data from the [Component
Ownership
Onboarding](https://docs.google.com/spreadsheets/d/1rPXYOHqKPYqMzpjuR25TRV7XV70Ml1Bo1MSmyK1hdG0/edit#gid=1041346743)
spreadsheet.

If a catalog-info.yaml file did not exist, we've created one. If it was
present, we've enhanced it by adding missing attributes from the
spreadsheet. In case of discrepancies, we prioritized the spreadsheet's
data. Please review the component information in this PR and make
changes if necessary.

## Next Steps
You are asked to review this PR as a code owner of the
`catalog-info.yaml` location. If you find that you are not the actual
owner of the component specified within this PR, please correct it in
this PR, and tag the actual owning team as a reviewer. Consider updating
the `CODEOWNERS` file to reflect the correct ownership for future
accuracy.

Once this PR is merged, the component's updated information will be
registered in Backstage.

## Notes
This PR is divided into two commits. The first commit standardizes the
format of `catalog-info.yaml`, whereas the second commit applies changes
based on data from the [Component Ownership
Onboarding](https://docs.google.com/spreadsheets/d/1rPXYOHqKPYqMzpjuR25TRV7XV70Ml1Bo1MSmyK1hdG0/edit#gid=1041346743)
spreadsheet. This should make reviewing changes to the catalog content
easier.

Additionally, all comments are removed during the format
standardization, due to limitations in the parser. We believe these
comments are mostly auto-generated, so their removal is not harmful.
However, please proceed with caution if you have left meaningful
comments in these files.