Fix missing entities in the spec.json
Together with Raktima we found an error in the name conflict resolution in the update-spec script. When in several specs there are entities with the same names, the script suppose to add a suffix to the second entity to resolve the conflict, but it didn't work. Search for "Deduplicate schema definitions" in update-spec.ts for details.

Beside the fix there is another issue in `updateDefinitionReferences` when the function modifies a reference name in an entity that has been resolved as similar to an existing once in the merged spec.